add default gkr config (changes from #58)
